add asw but...not good as much as thought

+++ Stereo_Matching/src/MatchingCost.java
@@ -1,5 +1,5 @@
 public interface MatchingCost {
-  public double matchingCost(int[][] left, int[][] right);
+  public double matchingCost(double[][] left, double[][] right);
 
   public double matchingCostLab(int[][] left, int[][] right, int[][] leftLab, int[][] rightLab);
 }
